iio: adc: adrv9002: move OF specific code to it's own file
OF specific code is big enough that it deserves to be in its own file
(note arv9002_of.c is approx 1K lines). This also "alleviates" the size
of the main driver.

While at it, bring some order to OF_ADRV9002_PINCTL(),
OF_ADRV9002_DGPIO() and OF_ADRV9002_FH(). These macros were making
assumptions on variables names (anc actually being present). While that
is ok for function scope macros, it is not pretty for global scopes ones.
Hence, just add all arguments we need to the macros.
